# frontend/openapi.json
{
  "components": {
    "headers": {
      "ETag": {
        "description": "Strong version used with If-Match.",
        "schema": {
          "pattern": "^\\\"[1-9][0-9]*\\\"$",
          "type": "string"
        }
      },
      "TraceId": {
        "description": "Request trace identifier.",
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "parameters": {
      "CsrfToken": {
        "description": "Token bound to the current opaque session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string"
        }
      },
...
      "IfMatch": {
        "description": "Numeric entity tag returned by the resource. Strong and weak HTTP forms are accepted.",
        "in": "header",
        "name": "If-Match",
        "required": true,
        "schema": {
          "pattern": "^(?:W/)?\\\"[1-9][0-9]*\\\"$",
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Conflict": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "The resource state conflicts with the request.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
      "Forbidden": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Permission, assignment or CSRF validation failed.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
      "PreconditionRequired": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "A valid If-Match header is required.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Request rejected.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "A valid opaque session cookie is required.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      }
    },
    "schemas": {
...
      "CreateScheduleRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ScheduleAction"
          },
          "enabled": {
            "default": true,
            "type": "boolean"
          },
          "instance_id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "trigger": {
            "$ref": "#/components/schemas/ScheduleTrigger"
          }
        },
        "required": [
          "instance_id",
          "name",
          "trigger",
          "action"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "trace_id"
        ],
        "type": "object"
      },
...
      "Schedule": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ScheduleAction"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "type": "string"
          },
          "last_job_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "last_run_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "next_run_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "requested_by": {
            "type": "string"
          },
          "trigger": {
            "$ref": "#/components/schemas/ScheduleTrigger"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "instance_id",
          "name",
          "trigger",
          "action",
          "enabled",
          "next_run_at",
          "last_run_at",
          "last_job_id",
          "version",
          "created_by",
          "requested_by",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "ScheduleAction": {
        "discriminator": {
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "start"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "stop"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "restart"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "backup"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "update"
              }
            },
            "required": [
              "kind"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "command": {
                "maxLength": 4096,
                "minLength": 1,
                "pattern": "^[^\\u0000\\r\\n]+$",
                "type": "string"
              },
              "kind": {
                "const": "console"
              }
            },
            "required": [
              "kind",
              "command"
            ],
            "type": "object"
          }
        ]
      },
      "ScheduleTrigger": {
        "discriminator": {
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "expression": {
                "description": "Six- or seven-field cron expression; each field is limited to 64 characters.",
                "maxLength": 454,
                "minLength": 11,
                "type": "string"
              },
              "kind": {
                "const": "cron"
              },
              "timezone": {
                "description": "IANA time-zone identifier.",
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "kind",
              "expression",
              "timezone"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "kind": {
                "const": "interval"
              },
              "seconds": {
                "maximum": 31536000,
                "minimum": 60,
                "type": "integer"
              }
            },
            "required": [
              "kind",
              "seconds"
            ],
            "type": "object"
          }
        ]
      },
...
      "SuccessResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateScheduleRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ScheduleAction"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "trigger": {
            "$ref": "#/components/schemas/ScheduleTrigger"
          }
        },
        "required": [
          "name",
          "trigger",
          "action",
          "enabled"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "Opaque HttpOnly SameSite=Strict session cookie. Secure is mandatory outside loopback development.",
        "in": "cookie",
        "name": "dmx_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-host game server management API. Secret values are write-only.",
    "title": "DmxServerManager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/schedules": {
      "get": {
        "operationId": "listSchedules",
        "parameters": [
          {
            "in": "query",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Schedules for the selected instance."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "createSchedule",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "Schedule created.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}": {
      "delete": {
        "operationId": "deleteSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Schedule removed."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "schedules"
        ]
      },
      "get": {
        "operationId": "getSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "Schedule.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "schedules"
        ]
      },
      "put": {
        "operationId": "updateSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "Schedule updated.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "schedules"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}